requirements-jobs.txt
sphinx
recommonmark
sphinx-autobuild

ipykernel
nbsphinx
pylint
autopep8
pytest
pytest-cov
pytest-xdist
pytest-sugar
tox
coveralls
pluggy<1,>=0.12.0
docstr-coverage
git+https://github.com/achillesrasquinha/bpyutils.git@develop#egg=bpyutils
Pillow
deepface
pyserial
pyaudio
pipenv
jedi<0.18
ipython
bumpversion
vulture
safety
pre-commit
twine

jupyter

Flask
